+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 # Unreleased
 - add DuckDB::PreparedStatement#prepare.
+- DuckDB::Connection#prepared_statement accepts block.
+    ```ruby
+    con.prepared_statement('SELECT * FROM table WHERE id = ?') do |stmt|
+      stmt.bind(1)
+      stmt.execute
+    end
+    ```
 
 # 1.1.2.0 - 2024-10-20
 - bump duckdb to 1.1.2.